Audio upload casi y falta arreglar delete publication

+++ routes/users_routes.py
@@ -10,12 +10,13 @@ from model.user_shemas import NewUser, User, UserProfile
 from model.lyrics_shemas import Lyrics
 from config.security import oauth2_scheme, guardar_log, SSH_USERNAME_RES, SSH_PASSWORD_RES, SSH_HOST_RES, \
     get_current_user, ACCESS_TOKEN_EXPIRE_MINUTES, create_access_token, get_user_id
-from config.db import users_collection, interactions_collection, get_database, lyrics_collection, follows_collection
+from config.db import users_collection, interactions_collection, get_database, lyrics_collection, follows_collection, post_collection
 from fastapi.security import OAuth2PasswordRequestForm
 from fastapi import File, HTTPException, Depends, UploadFile, status, APIRouter
 import paramiko
 from routes.follow_routes import get_followers, get_following
 from routes.posts_routes import list_user_publications
+import pymongo
 
 # Iniciar router
 router = APIRouter()
@@ -47,7 +48,7 @@ async def register(user: NewUser):
             raise HTTPException(status_code=500, detail="Error creating the folder on the remote server")
     return {"_id": str(result.inserted_id)}
 
-@router.delete("/delete-user")
+@router.delete("/delete")
 async def delete_user(current_user: NewUser = Depends(get_current_user), db=Depends(get_database)):
     if not current_user:
         raise HTTPException(
@@ -73,30 +74,42 @@ async def delete_user(current_user: NewUser = Depends(get_current_user), db=Depe
 
         # Obtener el ID del usuario
         user_id = await get_user_id(current_user.username)
+        if not user_id:
+            raise HTTPException(status_code=500, detail="User ID not found")
 
         # Obtener los IDs de los posts del usuario
-        user_posts = await lyrics_collection.find({"user_id": user_id}, {"_id": 1}).to_list(None)
+        user_posts = await post_collection.find({"user_id": ObjectId(user_id)}, {"_id": 1}).to_list(None)
         post_ids = [post["_id"] for post in user_posts]
-
+        #ELiminar los follows
+        await follows_collection.delete_many({"user_id_following": user_id})
+        await follows_collection.delete_many({"user_id_followed": user_id})
+        #Eliminar las letras
+        await lyrics_collection.delete_many({"user_id": user_id})
         # Eliminar todas las interacciones asociadas a los posts del usuario
         await interactions_collection.delete_many({"post_id": {"$in": post_ids}})
-
-        # Eliminar todas las interacciones hachas por el usuario
         await interactions_collection.delete_many({"user_id": user_id})
-
         # Eliminar todos los posts del usuario
-        await lyrics_collection.delete_many({"user_id": user_id})
-
+        await post_collection.delete_many({"user_id": user_id})
         # Eliminar al usuario de la colección de usuarios
-        await users_collection.delete_one({"_id": user_id})
+        delete_result = await users_collection.delete_one({"_id": ObjectId(user_id)})
+        if delete_result.deleted_count == 1:
+            print("User deleted successfully from database")
+        else:
+            print("User deletion failed or user not found in database")
 
+        return {"message": "User deleted successfully"}
+
+    except paramiko.AuthenticationException as e:
+        raise HTTPException(status_code=500, detail="SSH authentication failed")
     except paramiko.SSHException as e:
-        raise HTTPException(status_code=500, detail=f"SSH error: {str(e)}")
+        raise HTTPException(status_code=500, detail="SSH connection error")
+    except paramiko.ssh_exception.NoValidConnectionsError as e:
+        raise HTTPException(status_code=500, detail="No valid SSH connections available")
+    except pymongo.errors.PyMongoError as e:
+        raise HTTPException(status_code=500, detail="Database error: " + str(e))
     except Exception as e:
-        raise HTTPException(status_code=500, detail=f"An error occurred: {str(e)}")
-    
-    return {"message": "User deleted successfully"}
-   
+        raise HTTPException(status_code=500, detail="An error occurred: " + str(e))
+
 # Recoger datos del usuario actual
 @router.get("/users/me")
 async def read_users_me(
@@ -152,20 +165,36 @@ async def login_for_access_token(form_data: OAuth2PasswordRequestForm = Depends(
 async def get_saved_posts(current_user: NewUser = Depends(get_current_user), db=Depends(get_database)):
     user_id = await get_user_id(current_user.username)
     saved_posts = await interactions_collection.find({"user_id": user_id, "saved_date": {"$exists": True}}).to_list(None)
+    
+    # Convertir ObjectId a cadenas
+    for post in saved_posts:
+        post["_id"] = str(post["_id"])
+    
     return {"saved_posts": saved_posts}
 
-# Obtener todas las publicaciones likeadas por el usuario
+
 @router.get("/liked-posts")
 async def get_liked_posts(current_user: NewUser = Depends(get_current_user), db=Depends(get_database)):
     user_id = await get_user_id(current_user.username)
     liked_posts = await interactions_collection.find({"user_id": user_id, "like_date": {"$exists": True}}).to_list(None)
+    
+    # Convertir ObjectId a cadenas
+    for post in liked_posts:
+        post["_id"] = str(post["_id"])
+    
     return {"liked_posts": liked_posts}
 
 @router.get("/lyrics", response_model=List[Lyrics])
 async def get_user_lyrics(current_user: NewUser = Depends(get_current_user), db=Depends(get_database)):
     user_id = await get_user_id(current_user.username)
     user_lyrics = await lyrics_collection.find({"user_id": user_id}).to_list(None)
+    
+    # Convertir ObjectId a cadenas
+    for lyric in user_lyrics:
+        lyric["_id"] = str(lyric["_id"])
+    
     return user_lyrics
+
 
 
 @router.get("/profile/{user_id}", response_model=UserProfile)
@@ -174,7 +203,7 @@ async def get_user_lyrics(user_id: str, current_user: NewUser = Depends(get_curr
     if user_dict:
         userindb = User(**user_dict)
         user_id_following = await get_user_id(current_user.username)
-        if user_id_following:
+        if user_id_following==user_id:
             isFollowing = 0
         else:
